requirements.txt
pandas==1.1.5
psycopg2-binary==2.8.6
python-dotenv==0.15.0
